Fix UI JS error caused by empty image list
If there were no previously archived images in the UI's dropdown list, a JavaScript error occurred and prevented the rest of the JS from working (i.e. the `Archive!` button didn't work). This change checks to make sure there is a `name` value in the dropdown before attempting to load the properties associated with it.